Kamui's gender changeable, code gender swap bug fix
There was a bug where changing Kamui's gender in code would confuse the
program into thinking it wasn't changed and reset the swap button. Code
simplified to simply swap whatever is there with the opposite.

Kamui's portrait can now be swapped between male and female. As a
convenience, FEITS Exporter will change Kamui's portrait to match the
gender specified in a message's prefix/name, if there is one.